LayoutTests/fast/inline/inline-width-containing-collapsed-whitespace-and-image-in-float.html

<!DOCTYPE html>
<style>
body { font-family: sans-serif, Arial; }
li { float: left; list-style: none; }
</style>
<ul>
  <li>
    <a href="#">
      <img src="./resources/checker.png">
    </a>
  </li>
</ul>
<script src="../../resources/js-test.js"></script>
<script>
description("Test if width of inline box including a image and whitespaces is same with the included image width.");
window.onload = function() {
  shouldBeEqualToNumber("document.querySelector('li').getBoundingClientRect().width", 32);
}
</script>
